Type check `@abi` for compatibility
This commit compares the decl inside the `@abi` attribute to the decl it’s attached to, diagnosing ABI-incompatible differences or attributes that don’t need to be specified.

It includes logic which ensures that some of the fundamental properties of a declaration, such as its availability, access control, and @objc-ness, are shared between an ABI-only declaration and its counterpart. (Specifically, the ABI-only declaration delegates the computation to its counterpart.) This allows these attributes to be omitted from the `@abi` attribute instead of forcing developers to repeat themselves.