CMake: Implement dispatch search on Linux and Windows
Adding implementations for finding libdispatch installed on Windows and
Linux. The module uses a `Swift_SDKROOT` variable as a hint on where to
look, as well as the SDKROOT environment variable on Windows to match
the behavior of the compiler. This allows re-using an existing
installation of corelibs-libdispatch without needing to rebuild it to
get access to the dispatchConfig.cmake file.