[cxx-interop] Remove code from SwiftifyImport that drops _const-suffix (#79756)
This was introduced in 140552054c45fc4e2795aed96e39b63c04052288 in an
attempt to be accommodate existing interfaces with the _const scheme,
but we shouldn't need to support such textual interfaces anyway since
the _const suffix only appears in C++ code (which shouldn't be part of
a resilient interface).

For newly generated interfaces, dropQualifierSuffix doesn't do anything
since the code that generates the suffix is now gone.

rdar://143769901