Build Preset: Bootstrap Stage-0 Toolchain
This preset builds the compiler and standard library on a system without
an existing Swift.

Note, there are many issues with it and it won't be able to pass tests.
It is an intermediate product that will need to build a compiler with
the Swift sources. The stage-0 compiler won't be able to build the
driver as the driver depends on Foundation, which requires macro support
and the stage-0 compiler does not have macro support. It will need to
build swiftsyntax though to give the stage-1 compiler macro support.
The stage-1 compiler then can build a stage-2 toolchain with most of the
libraries and runtimes. That stage-2 toolchain should then be capable of
building the full nightly toolchain package.