Index synthesized wrapper refs by substituting the Clang decl.
The goal of this change is replace the references to synthesized declarations, whose definitions are never contained in the indexstore, with references to the underlying declarations which are contained in the indexstore. The current behavior prevents code navigation from the references to the declaration.

Reference: https://github.com/swiftlang/swift/issues/79912

- I added the EnumElementDecl generated from the underlying Clang EnumConstantDecl to the imported decl cache for ClangImporter.
- I took the effective Clang node and corresponding imported decl to substitute for "synthesized wrapper decls".

This has a weird outcome, where a reference to the `Code` enum inside of the synthesized struct can emit 2 references to the underlying enum. There's an example in the test case, where `TestError.Code` is referenced.

Fixes 79912

adding emitDataForSwiftSerializedModule emits the decls but:
- The locations are all 0:0
- The decls are emitted in the file "module.modulemap